requirements.txt
# Core dependencies
pypdf>=3.17.0
pdfplumber>=0.10.0
langchain-text-splitters>=0.2.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
numpy>=1.24.0
python-dotenv>=1.2.1
streamlit>=1.52.1

# LLM providers
anthropic>=0.75.0
openai>=2.9.0
tiktoken>=0.12.0

# API later
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.12.0
requests>=2.32.5
python-multipart>=0.0.20
